g3proxy: switch to IdleWheel for IDLE check

+++ g3proxy/src/auth/user.rs
@@ -67,7 +67,7 @@ pub(crate) struct User {
 
 impl User {
     #[inline]
-    pub(crate) fn task_max_idle_count(&self) -> i32 {
+    pub(crate) fn task_max_idle_count(&self) -> usize {
         self.config.task_idle_max_count
     }
